+++ gems/aws-sdk-resources/lib/aws-sdk-resources.rb
@@ -223,6 +223,7 @@ module Aws
   autoload :MediaConvert, 'aws-sdk-mediaconvert'
   autoload :MediaLive, 'aws-sdk-medialive'
   autoload :MediaPackage, 'aws-sdk-mediapackage'
+  autoload :MediaPackageV2, 'aws-sdk-mediapackagev2'
   autoload :MediaPackageVod, 'aws-sdk-mediapackagevod'
   autoload :MediaStore, 'aws-sdk-mediastore'
   autoload :MediaStoreData, 'aws-sdk-mediastoredata'