+++ gems/aws-sdk-sesv2/lib/aws-sdk-sesv2/endpoints.rb
@@ -935,6 +935,20 @@ module Aws::SESV2
       end
     end
 
+    class PutDedicatedIpPoolScalingAttributes
+      def self.build(context)
+        unless context.config.regional_endpoint
+          endpoint = context.config.endpoint.to_s
+        end
+        Aws::SESV2::EndpointParameters.new(
+          region: context.config.region,
+          use_dual_stack: context.config.use_dualstack_endpoint,
+          use_fips: context.config.use_fips_endpoint,
+          endpoint: endpoint,
+        )
+      end
+    end
+
     class PutDedicatedIpWarmupAttributes
       def self.build(context)
         unless context.config.regional_endpoint