+++ gems/aws-sdk-backup/CHANGELOG.md
@@ -1,6 +1,11 @@
 Unreleased Changes
 ------------------
 
+1.50.0 (2023-05-19)
+------------------
+
+* Feature - Add  ResourceArn, ResourceType, and BackupVaultName to ListRecoveryPointsByLegalHold API response.
+
 1.49.0 (2023-02-08)
 ------------------