+++ gems/aws-sdk-connectcases/CHANGELOG.md
@@ -1,6 +1,11 @@
 Unreleased Changes
 ------------------
 
+1.5.0 (2023-05-19)
+------------------
+
+* Feature - This release adds the ability to create fields with type Url through the CreateField API. For more information see https://docs.aws.amazon.com/cases/latest/APIReference/Welcome.html
+
 1.4.0 (2023-02-24)
 ------------------